"with" will be a Python keyword starting in Python 2.6 (and Python 2.5 already emits warnings if "with" is used as an identifier.  So...
Do not use "with" as an identifier within Gnuplot.py (including the
test and demo scripts).  Rename the PlotItem "with" parameter to
"with_".  However, for backwards compatibility (and through the magic
of Python varargs processing), accept "with" as an alternative
spelling of this option.

This version runs on my computer, but I only have Python 2.4
installed, so I can't really test it.


git-svn-id: svn://svn.code.sf.net/p/gnuplot-py/code/trunk@299 67c13c2a-ce2d-4362-9fbd-a73952e22801