Fix handling of float64 data.
The old code attempted to handle Python data as float32 (to save
memory).  But contrary to the comment, it was coercing float64 numpy
arrays to float32.  (Perhaps this is a difference between numpy and
Numeric?)  That was never the intention.

The new version avoids this coersion, with the side-effect that Python
lists are converted to float64 numpy arrays.

Patch by: xaverxn (see tracker patch with ID 2671685)

> Gnuplot.py float64 handling is broken (gets downcast to float32, see
> gnuplot-py-users mailing list in the beginning of march 09).  It is
> essentially my last proposal on the list, which wasn't discussed or
> commented on any further.
>
> This patch changes the file utils.py so it coverts python native types
> to numpy arrays and (up)casts everything but float64 to float32.

git-svn-id: svn://svn.code.sf.net/p/gnuplot-py/code/trunk@316 67c13c2a-ce2d-4362-9fbd-a73952e22801